Add support for SSH authentication
When attempting to authenticate to an SSH server using the `enable-ssh-support`
feature of `gpg-agent`, the `pinentry-touchid` program would crash and an error
message like this would be printed by the `gpg-agent` program to the console:

    panic: runtime error: index out of range [1] with length 0

    goroutine 1 [running]:
    main.GetPIN.func1({{0xc00006e0e0, 0x61}, {0xc00001815a, 0xb}, {0x0, 0x0}, {0x0, 0x0}, {0x0, 0x0}, ...})
	    .../pinentry-touchid/main.go:246 +0xc13

The root cause of this error is that `pinentry-touchid` assumes that the
passphrase prompt for a key will always contain an email address, e.g.,:

    Please enter the passphrase to unlock the OpenPGP secret key:
    "Cathy J. Fitzpatrick <cathy@cathyjf.com>"
    4096-bit RSA key, ID ...,
    created ....

However, the passphrase prompt for an SSH key is different. It looks like this:

    Please enter the passphrase for the ssh key
      SHA256:<hash>

where <hash> appears to be the base-64 encoded version of an SHA256 hash for
the key in question. Notably, this passphrase prompt does not contain an email
address, so `pinentry-touchid` is unable to parse it.

This commit fixes this problem by adding the ability for `pinentry-touchid` to
parse the passphrase prompt for SSH keys. SSH key passphrases will be saved
in keychain, similar to other keys, with the label being "SSH (<hash>)" where
<hash> is the same base-64 encoded hash found in the passphrase prompt.